INVOICE_MANAGEMENT/requirements.txt
mysql_connector_repackaged==0.3.1
PyPDF2==3.0.1
python-dotenv==1.2.1
